style: lowerCamelCase for action and combine ServiceMethod.ts and ServiceMethodTypes.ts
