+++ src/actions/career/companies/create.ts
@@ -1,5 +1,5 @@
 'use server'
-import { Action } from '@/actions/Action'
+import { action } from '@/actions/action'
 import { createCompany } from '@/services/career/companies/create'
 
-export const createCompanyAction = Action(createCompany)
+export const createCompanyAction = action(createCompany)
