+++ src/actions/career/jobAds/read.ts
@@ -1,7 +1,7 @@
 'use server'
-import { Action } from '@/actions/Action'
+import { action } from '@/actions/action'
 import { readActiveJobAds, readInactiveJobAdsPage, readJobAd } from '@/services/career/jobAds/read'
 
-export const readJobAdAction = Action(readJobAd)
-export const readActiveJobAdsAction = Action(readActiveJobAds)
-export const readInactiveJobAdsPageAction = Action(readInactiveJobAdsPage)
+export const readJobAdAction = action(readJobAd)
+export const readActiveJobAdsAction = action(readActiveJobAds)
+export const readInactiveJobAdsPageAction = action(readInactiveJobAdsPage)
