+++ src/actions/dots/read.ts
@@ -1,7 +1,7 @@
 'use server'
-import { Action } from '@/actions/Action'
+import { action } from '@/actions/action'
 import { readDotsPage, readDotWrappersForUser } from '@/services/dots/read'
 
-export const readDotPageAction = Action(readDotsPage)
+export const readDotPageAction = action(readDotsPage)
 
-export const readDotWrappersForUserAction = Action(readDotWrappersForUser)
+export const readDotWrappersForUserAction = action(readDotWrappersForUser)
