+++ src/actions/events/tags/update.ts
@@ -1,5 +1,5 @@
 'use server'
-import { Action } from '@/actions/Action'
+import { action } from '@/actions/action'
 import { updateEventTag } from '@/services/events/tags/update'
 
-export const updateEventTagAction = Action(updateEventTag)
+export const updateEventTagAction = action(updateEventTag)
