+++ src/actions/groups/interestGroups/read.ts
@@ -1,5 +1,5 @@
 'use server'
-import { Action } from '@/actions/Action'
+import { action } from '@/actions/action'
 import { readAllInterestGroups } from '@/services/groups/interestGroups/read'
 
-export const readInterestGroupsAction = Action(readAllInterestGroups)
+export const readInterestGroupsAction = action(readAllInterestGroups)
